requirements.txt
pandas
numpy
matplotlib
seaborn
chainlit
Pillow
google-generativeai